Move all transformers together
