+++ static/app/views/dashboards/utils/transformEventsResponseToSeries.tsx
@@ -6,7 +6,8 @@ import type {
 } from 'sentry/types/organization';
 
 import type {WidgetQuery} from '../types';
-import {transformSeries} from '../widgetCard/widgetQueries';
+
+import {transformEventsStatsToSeries} from './transformSeries';
 
 type SeriesWithOrdering = [order: number, series: Series];
 
@@ -27,7 +28,7 @@ export function transformEventsResponseToSeries(
     const field = widgetQuery.aggregates[0]!;
     const prefixedName = queryAlias ? `${queryAlias} : ${field}` : field;
 
-    seriesWithOrdering.push([0, transformSeries(data, prefixedName, field)]);
+    seriesWithOrdering.push([0, transformEventsStatsToSeries(data, prefixedName, field)]);
   } else if (isMultiSeriesEventsStats(data)) {
     Object.keys(data).forEach(seriesName => {
       const seriesData = data[seriesName]!;
@@ -35,7 +36,7 @@ export function transformEventsResponseToSeries(
 
       seriesWithOrdering.push([
         seriesData.order ?? 0,
-        transformSeries(seriesData, prefixedName, seriesName),
+        transformEventsStatsToSeries(seriesData, prefixedName, seriesName),
       ]);
     });
   } else if (isGroupedMultiSeriesEventsStats(data)) {
@@ -55,7 +56,7 @@ export function transformEventsResponseToSeries(
 
         seriesWithOrdering.push([
           (groupData.order as unknown as number) ?? 0,
-          transformSeries(seriesData, prefixedName, seriesName),
+          transformEventsStatsToSeries(seriesData, prefixedName, seriesName),
         ]);
       });
     });
